% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nano_obj.R
\name{create_nano_obj}
\alias{create_nano_obj}
\title{Creates a nano object}
\usage{
create_nano_obj(sample.data, id.data, count.data, row.data)
}
\arguments{
\item{sample.data}{A data frame of sample information such as ROI names, Area, Muclei count etc. The rownames are the attributes and the column names are the ROI identifier}

\item{id.data}{A data.frame with information about the ROI such as Condition (Infected or Uninfected)}

\item{count.data}{The count matrix. The rownames should match the rownames of row.data and column names should match the rownames of id.data}

\item{row.data}{A data.frame containing information about the features (RNA or Protein). CodeClass column should be one of the columns.}
}
\description{
Creates a S3 object of all the information about target features and samples. This object is used for the other functions in this package. An item on this object can be access as slot(nano_obj, item_name)
}
\details{
From the matrices of count data, sample information, and target information this function creates a nano object
}
\author{
Subrata Paul
}
