% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{Plots principal components}
\usage{
pca_plot(
  nano_obj,
  slot_id = "count",
  array_name = NULL,
  features = "Endogenous",
  x_index = 1,
  y_index = 2,
  color_by = NULL
)
}
\arguments{
\item{nano_obj}{A nano object}

\item{slot_id}{Name of the slot to use for PCA. One of count and count_norm for raw and normalized count matrix, respectively.}

\item{array_name}{If count_norm slot is used, name of the array within the slot. The name is same as given in the normalization step}

\item{features}{One of "Endogenous", "Negative" or "Control" to select the features within these CodeClass. Also a vector of selected features can be given containing more than one features.}

\item{x_index}{Index of PC on the x axis (default 1)}

\item{y_index}{Index of PC on the y axis (default 2)}

\item{color_by}{Column of id data to use to color the points}
}
\description{
Plots principal components
}
