% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pairs.R
\name{qc_pairs}
\alias{qc_pairs}
\title{Pairs plot of set of features}
\usage{
qc_pairs(
  nano_obj,
  target,
  category_col = "Target_Group",
  color_by,
  shape_by = NA
)
}
\arguments{
\item{nano_obj}{A nano object to plot}

\item{target}{Character string identifying the set of proteins/RNA.}

\item{category_col}{Column to used to identify the target set. By default 'Target_Group'. The count matrix will be subsetted by the rows that contains "target" in the "category_col".}

\item{color_by}{Column in the id data to be used for coloring the points}

\item{shape_by}{Column in the id data to determine the shape of the points. By default no shape is used.}
}
\description{
Depending on the selected set of RNA or Proteins it will create a pairs plot of the Proteins within the set.
}
