% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_scatter_between.R
\name{qc_scatter_between}
\alias{qc_scatter_between}
\title{Scatter Plot of Housekeeping and Background}
\usage{
qc_scatter_between(nano_obj, color_by, shape_by = NA)
}
\arguments{
\item{nano_obj}{A nano object to plot}

\item{color_by}{Column name os id data to be used to color the points}

\item{shape_by}{Column name of id data to be used to identify the points}
}
\description{
Scatter Plot of Housekeeping and Background
}
