% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLEplot.R
\name{RLEplot}
\alias{RLEplot}
\title{Create RLE plot}
\usage{
RLEplot(nano_obj, count_slot = "count", assay_name = NULL, color_by = NULL)
}
\arguments{
\item{nano_obj}{A Nano object}

\item{count_slot}{Slot to grab the count matrix}

\item{assay_name}{If count matrix is from count_norm there are matrices with different name. Provide name of the count matrix to use}

\item{color_by}{Color the boxplot by any column on the id data}
}
\description{
Create RLE plot
}
\author{
Subrata Paul
}
