% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_to_bg_ratio.R
\name{signal_to_bg_ratio}
\alias{signal_to_bg_ratio}
\title{Signal to Background Ratio}
\usage{
signal_to_bg_ratio(nano_obj, neg_controls = NULL)
}
\arguments{
\item{nano_obj}{A nano object}

\item{neg_controls}{A vector of negetive controls if not all negative controls contains string 'IgG'}
}
\description{
The IgG genes are used as background in the nanostring experiment. Signal to background ratio determines how well the target expression were counted compared to background noise.
This function looks for 'IgG' sring on the gene/protein names to identify the negative controls. For other situation a vector of negetive controls can be supplied.
The signal to background ratio is calculated as the ratio of counts divided by the geometric mean of negative controls.
The function gives a boxplot with individual points of the signal to background ratio.
}
