% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{umap_plot}
\alias{umap_plot}
\title{Plot UMAP}
\usage{
umap_plot(nano_obj, slot_id = "count", array_name = NA, color_by = "Status")
}
\arguments{
\item{nano_obj}{A nano object}

\item{slot_id}{Either count or count_norm to use raw count or normalize count, respectively.}

\item{array_name}{The array name of the normalize count}

\item{color_by}{Column of id data used for coloring the points}
}
\description{
Plot UMAP
}
